discord_bot/requirements.txt
discord.py
aiohttp
python-dotenv
watchfiles